Second pass at core/collate.py refactor: Move the file to pypeit/collate.py
