+++ pypeit/collate.py
